Add mix LED mode.

+++ simdelay.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void Delay100Us(uint8_t x);
-void DelayMs(uint8_t x);
+void DelayMs(uint16_t x);
 
 #endif
